requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Google ADK + Gemini (use google-genai, NOT google-generativeai — they conflict)
google-adk>=0.4.0
google-genai>=1.0.0
google-cloud-firestore>=2.19.0

# Session storage
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.30.0

# Scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Utilities
python-dotenv>=1.0.0

# Auth
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
stripe>=10.0.0

# Email
aiosmtplib>=3.0.0
certifi>=2024.0.0
pytz>=2024.0
google-cloud-texttospeech
pydub
requests